Add assertion-verifier chart and app configmap flag
Created assertion-verifier chart and added app configmap flag.
When set to true will use the helm chart application.yml.
That is the default now as it is not changing often. Re-packaged
charts, added proper metrics paths, tuned health-checks, and
changed ports to the application ports used in a local environment.

Testing:
 - Installed model-builder and assertion-detector charts on the
   asserts-dev cluster.
 - Verified rules and application configs were correct.

TODO:
 - Add assertion-verifier ci/cd in adi project. It will be required
   to use this chart as there are currently no images for it in ecr.